WIP: Introduce BindingInstanceIdentifier
binding.InstanceIdentifier does not really fulfill the contract we need
-- it cannot be constructed for all valid 'type instance-identifier'
values. Furthermore it has a 'wildcard' overload, which we do not want.

This patch introduces BindingInstanceIdentifier as the core type
mapping, with InstanceIdentifier's place as the type mapping. Its role
in 'points to a DataObject' role is taken over by DataObjectReference,
which does not allow wildcards, but otherwise mirrors its behaviour.

// FIXME: issue

Change-Id: I88dc31cb5285bfc9af01c28227cb4ed29b6bfe84
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>