Introduce DataObject{Identifier,Reference}
InstanceIdentifier has way cluttered API, combining multiple aspects,
without giving useful guarantees.

We introduce DataObjectReference, which captures the addressing
capabilities of an InstanceIdentifier.

We also introduce DataObjectIdentifier, which guarantees isExact() is
true -- and forms the baseline BindingInstanceIdentifier.

Finally, we also the baseline 'WithKey' trait, so it is readily
available as the impact of the target type.

JIRA: YANGTOOLS-1577
Change-Id: Ieb87705f65c2032fdb499cdd74b55c5994718566
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>