WIP: Seal KeyAware
We are using KeyAware with conjunction with DataObject, leaving
generated code to connect the two. This is unfortunate, as that
conjuction is really a kind of DataObject, similar to ChildOf.

Introduce a dedicated type, EntryObject, which combines DataObject and
KeyAware, providing a nice marker for generated interfaces. This also
allows us to further specialize KeyAware and seal its class hierarchy,
reducing potential confusion.

// FIXME: add an issue

Change-Id: I11a1d1c678c5e7f4397d1a2ddf420067559b2dbd
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>