Rework Grouping mapping
Having interfaces generated for grouping extend DataObject brings in
a lot of chaos, as it implies Grouping is addressable, which it is not.

We have a lower-impact contract, DataContainer, which gives similar
facilities, but is does not imply the interface can be used in
DataObjectStep.

Introduce binding.Grouping as a DataContainer specialization and switch
interfaces generated for groupings to use that.

This also flushes out a problem with addressing, as builderOfInherited()
requires a DataObject in its signature. Fix this by making DataRoot a
DataContainer contract as well.

JIRA: YANGTOOLS-1577
Change-Id: I9c63400281098a25974feb46ae33101793f3248b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>