Add PropertyStep
PropertyStep identifies a terminal node in the DataContainer tree
structure -- i.e. something which is not a DataContainer. An
ExactPropertyStep is a valid BindingInstanceIdentifier.Step.

This also adds a PropertyIdentifier as a BindingInstanceIdentifier
subclass -- combinding a DataObjectIdentifier and a ExactPropertyStep.

This ends up almost completing the binding-spec level of changes, except
it highlights the problem that an OpaqueObject is a DataContainer: and
that should not be the case.

JIRA: YANGTOOLS-1577
Change-Id: Ia45cc327b4767d8317468ddef7b4c06f497c3d04
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>