WIP: Introduce DataTreeCandidate.CandidateNode
DataTreeCandidateNode's design dates back to Java 7 days, which shows in
its layout -- users need to understand behaviours implied by
ModificationType and do not get much other help.

This patch introduces CandidateNode, which is a sealed interface
exposing the bare basics and a number of non-sealed specializations to
handle all the different cases that may occur.

This allows users to use enhanced switch statements to differentiate
between the cases, working quite nicely with Java 21.

We also re-structure the SPI implementations to reduce confusion and
prepare for interop.

// FIXME: issue

Change-Id: I9f2cb5365f527acd9ecb1b684de2f09abe9083db
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>