WIP: Update InstanceNotification mapping
InstanceNotification is capturing its parent, which can reasonably be
Grouping, which is not a DataObject.

The situation is not dissimilar to that of RpcInput/RpcOutput for
actions. There we generate a data anchor at the place of definition for
wide reuse -- but generate an interface for each instantiation in
non-grouping.

In this case, though, we cannot generate a DataObject, as we are
inherently tied to an instance identifier.

We therefore go the 'like a grouping' route, generating a reusable
interface, which is then inherited by all concrete instantiations.

// FIXME: tests for binding-data-codec-dynamic (should work, right?)

JIRA: YANGTOOLS-1577
Change-Id: I54383bc2fe1d060d20911861297ec59d49ddbfc7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>