Do not generate QNAME constaints for Groupings
QNAME constant is not really needed, as there is nothing on data
addressing side to do. Futhermore, since static members are 'inherited',
this ends up leading to some augmentations having a QNAME member.

Do not add the constant in GroupingGenerator and adjust tests to pick up
the namespace they need from a stable location.

JIRA: YANGTOOLS-1577
Change-Id: Ic03c0d6980317bb8042b482efd03fb9befb00283
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>