Add (Dynamic)BindingDataCodec et al.
We have BindingDOMCodecServices, which provides:
- a consistent set of binding-data-codec-api services
- a BindingRuntimeContext
- (indirectly) a EffectiveModelContext

The first item is a binding.data.codec.api concern, which is addressed
by exposing BindingDataCodec, which exposes the individual services. It
also takes care of the third item via an explicit modelContext() method,
as yang.model.api is implied in the services we provide.

The second item is the assumption of a dynamic runtime, where
BindingRuntimeContext is readily available. This is exposed via
binding.data.codec.dynamic.DynamicBindingDataCodec.

BindingDOMCodecServices are deprecated in favor of using one of these
interfaces.

We also introduce binding.data.codec.dynamic.BindingDataCodecFactory,
which supersedes binding.data.codec.spi.BindingDOMCodecFactory, as this
is an assumption of dynamism.

binding-data-codec-osgi now explicitly requires
binding-data-codec-dynamic, which is appropriate w.r.t. to assumption of
BindingRuntimeContext being passed around.

JIRA: YANGTOOLS-1638
Change-Id: I7d44063b42c3cddaa39c1c7a8aa28e035bb130a3
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>