Refactor SubstatementValidator error reporting
This teaches the SubstatementValidator to report multiple errors, using
Throwable.addSuppressed() chaining.

In order to make this work efficiently, we also turn the matching
algorithm around, with four distinct steps:

1. collect statement counts map, as we've done before
2. iterate through mandatory statements, removing from the counts map
   as we go
3. iterate through remainder of counts map
4. throw an exception

The code is refactored to allow for easy build up of reported
SourceException: we start at null, then we allocate a plain exception
and then then we use addSuppressed(), which is 5-line logic in two
copies.

We also improve reporting order:
- any missing mandatory substatement is reported in the Builder invocation
  order
- any non-mandatory is reported in input encounter order

JIRA: YANGTOOLS-1758
Change-Id: Ibcd62b7e88e518ca30b1c27323a3388bb857cbb5
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>