Introduce InMemoryDataTreeModification.State
InMemoryDataTreeModification is a nexus of multiple code paths, each
accessing differing aspects of a ModifiedNode.

This patch encapsulates rootNode in a State object, so that a state
check is required before we act on the rootNode.

We initially have only the states mirroring the
STATE_{OPEN,SEALING,SEALED} constants, but more states are expected to
be introduced by follow-up patches.

toString() method is adjusted to dump state, but states no longer dump
the entirety of the ModifiedNode, to prevent OOMs when toString() is
called on large modifications as part of logging/error reporting.

JIRA: YANGTOOLS-1651
Change-Id: I30a83b1944a3e27f56b243e6c64d4f8fa1b08d25
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>