WIP: Improve BindingClassLoader.create(...)
We have an mandatory and an optional part here. Let's use a Builder with
java.nio.Path internally.

This ends up making BindingCodecContext loading a tad more obvious, as
we configure a JVM-wide Builder and then just call build() -- amortizing
internal loading stuff.

Also improve forward compatibility: when we run with Java 24, we assume
AccessController indirection is a no-op as per JEP-486 and do not touch
it -- thus making the code immune to AccessController not being
available in future versions.

// FIXME: add an issue

Change-Id: Ifc63b1fe4d85f8aef1d0d43f6535e849bebb9bf4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>