Synchronize InMemoryDataTreeModification transitions
The three methods marked here all mutate ModifiedNode, with
newModification() happening concurrently with validate() and prepare().

This patch introduces a heavy-weight guard to prevent concurrent access,
so we serialize access between the threads.

It is not a complete solution, as newModification() can still
interfere wreck state inside a prepare()d candidate -- and we will deal
with that in a follow-up patch.

JIRA: YANGTOOLS-1651
Change-Id: If2132b3a6ac51c72dd8380b2ffdfb49158f023e8
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>