Introduce InMemoryDataTreeModification.Defunct
We use 'Sealing' as a transitional state indicating not-Open, but do not
update it if something bad happens -- leaving no evidence of the
failure for future reference.

Introduce a Defunct state, which carries the evidence of the failure
along with the name of the thread that incurred it.

Since this is the second singleton State object, introduce
SingletonState to hold the common bits.

JIRA: YANGTOOLS-1651
Change-Id: Icdda877c576611ed86391cfe149c5cc385061830
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit d1200fb3306ccdf7f43b5452c72aea0e8401b733)