Add DataObjectReference.firstKeyOf()
InstanceIdentifier class has been mark as deprecated since 14.0.0.
A .toLegacy() method has been introduced to
DataObjectReference class in order to ease the transition.

There is a glaring omission, which is forcing downstream users to bounce
DataObjectReferences through .toLegacy(): firstKeyOf().

Introduce DataObjectReference.firstKeyOf() as the generalization of
InstanceIdentifier.firstKeyOf(), with the implementation residing in
AbstractDataObjectReference.

While we are at it, we improve the documentation a bit and improve
argument checks, so as not to get defeated by raw type overrides.

JIRA: YANGTOOLS-1656
Change-Id: Ibea2e8eec664927f652f963fda2ce06d738941c4
Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>