Convert yang-data-util to bnd-parent
The first part of this is obvious -- change pom.xml, update
module-info.java, add package-info.java's to hold @Export.

The second part is less obvious: since we now explicitly control
exports, we can from 'impl' from 'impl.context'. This is a safe
operation, as we do not expose this package in JPMS nor in OSGi.

Change-Id: Iaea72ca0db8dc9deae2538e109e5b6668dcbe4b7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>