Add DataObjectIdentifier.Builder.toReferenceBuilder()
We have a use case where we start off with a DataObjectIdentifier, but
want to create a wildcard reference to something underneath it.
The only way to achieve this is to bridge to InstanceIdentifier's
builder -- which is cumbersome and not straightforward.

This patch adds a method that makes it much easier: now we can switch to
lenient DataObjectReferenceBuilder with a simple method.

JIRA: YANGTOOLS-1661
Change-Id: I9932251e9faf80059fe043eaed491019c514a767
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>