Add DataObjectIdentifier.trimTo()
DataObjectIdentifier is hierarchical, so there is a use case calling for
efficient trimming of these to their ancestor -- as evidenced by
InstanceIdentifier.firstIdentifierOf().

This patch adds the equivalent method, DataObjectIdentifier.tryTrimTo(),
which returns a nullable. We then use it to build a convenient
DataObjectIdentifier.trimTo(), which returns a non-null or throws an
unchecked exception.

JIRA: YANGTOOLS-1676
Change-Id: Ie5d356adb22a8bf22e748e1de72bd50685af6891
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>