Simplify OSGiBindingDOMCodec
We have constructor injection available, hence we can eliminate the
dance around active/inactive instances, as we simply get our basics
in early.

While we are here, we also improve shutdown a bit by not instantiating
the Map's values, but using a forEach() instead. We also clear the map
to make sure we do not retain any references.

JIRA: YANGTOOLS-1679
Change-Id: I65838fbf407b5c9d44059e636e94bb09e65b5428
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>