Separate out yang.parser.antlr.YangTextParser
The translation of YangTextSource to IRStatement (and corresponding
YangStatementStreamSource/YangIRSource) is scattered across two packages
and 6 classes.

This patch creates YangTextParser.parseToIR(YangTextSource) as the
common method. Since this would be a lonely static class, we also
integrate YangErrorListener in a manner similar to
SourceExceptionParser.

This allows us to:
- hide IRSupport, removing one of its methods
- eliminate CompactYangStatementLexer, as it is now a useless class

JIRA: YANGTOOLS-1709
Change-Id: Icf764f312463fd2eb9959ebd9b4a86ba52e2181d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>