Merge IfFeatureStatementSupport
The two supports differ only in which IfFeatureArgumentParser they
invoke. Flatten that dispatch into a field in a single final class.

That eliminates the need for a dedicated
yang.parser.rfc7950.stmt.if_feature package, which is thus eliminated.

We also take the time to use local variable type inference to provide
a bit of modernization.

JIRA: YANGTOOLS-1710
Change-Id: Iaaaf154a6074c3e691ae358387b6c5db7ae9e08b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>