Clean up 'deviate' statement support
As it turns out DeviateStatementRFC7950Support overrides only
SubstatementValidator, and only for two cases.

Use instance fields to carry which validators should be used and squash
the three classes into one -- eliminating a dedicated package in the
process of doing so.

While we are refactoring, also use local variable type inference.

JIRA: YANGTOOLS-1712
Change-Id: I56100fa1c9d32eabbdc0fa3a66df342c0290a3c4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>