Split out ReferenceDataTreeFactory
We have a normal module and sealed classes available. Split out
ReferenceDataTreeFactory, so that we can hide annotations not
interesting to users.

This allows us to return package-internal types as the result of the
methods.

While we are at it, migrate test suite to use the in-package factory and
do some amount of maintenance on the test suite.

JIRA: YANGTOOLS-1724
Change-Id: I82ee7284ff6235a3aeaf351cd8fb9f69dbfcf539
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>