Refactor ClassGeneratorBridge.BridgeProvider
BridgeProvider is needlessly tied to ClassGenerator, having a generic
arugment and forcing CodecDataObjectGenerator to have two subclasses.

Factor out GetterGenerator as the baseline class to generate getters,
with two specializations dealing with the particulars.

This eliminates the need to a have generic argument in BridgeProvider,
and relocating the loading customization to CodecDataObjectGenerator.

This provides two new top-level GetterGenerator specializations, which
not only can provide the current services without being tied to
CodecDataObjectGenerator, but also provide natural hosts for the
bytecode implementations they use.

JIRA: YANGTOOLS-1740
Change-Id: I2495990ebea8b2ae06cba2909eb9fd677bd6640d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>