Daggerize *-parser-support
This is a mass update of YANG parser extensions to expose Dagger
modules by default and deprecate current @Inject-based extensions.

We are introducing XXX.dagger package with one or more XXXModules, which
are not exported to OSGi.

Existing @Inject classes are refactored to delegate to the hidden
implementation extention.

JIRA: YANGTOOLS-1727
Change-Id: Ia00223ac68760d691e1a9b023d1116e0c921de1c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>