Use ParseException in ArgumentContextUtils
We are using StatementSourceReference only to construct SourceException,
which has the unfortunate effect of requiring YangIRSourceInfoExtractor
to instantiate the reference for each examined statement.

Switch to using ParseException, which does not need to reference and
convert it to SourceException in callers.

This eliminates the need to construct StatementSourceReference and
disconnects ArgumentContextUtils from yang-parser-spi constructs.

JIRA: YANGTOOLS-1748
Change-Id: I27c341cae0ad41da0b4b4c6517fb9ab2950637ff
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>