Add YinDomSource.SourceRefProvider
The parser reactor needs the ability to acquire
StatementSourceReference, which is something we are keeping private in
the implementation -- which in turn renders YinDomSource rather
inoperable.

Retrofit SourceRefProvider, which provides the reference based on
org.w3c.dom.Element and rehost the parsing bits (and our default
implementation) to model.spi.source.

This allows YinDomSource to become SourceInfo.Extractor by delegating
directly to YinDomSourceInfoExtractor.

JIRA: YANGTOOLS-1747
Change-Id: Ic53adf2493a862186b3dd66ef31fedd29ac6c0be
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>