Rehost RootStmtContext.definingModule()
Move the implementation to RootStatementContext. While this seems to
increase the model.api.stmt coupling, that is not the case -- it has
been there transitively via the use of StmtContextUtils.

Also switch from IllegalArgumentException, which made sense in
StmtContextUtils, to VerifyException, which more accurately models what
is going on.

JIRA: YANGTOOLS-1753
Change-Id: I2bc1b9fe8d27ea3f24bc286ed4ce05628ab453e0
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>